+++ contracts/commons/GlacisCommons.sol
@@ -20,7 +20,7 @@ contract GlacisCommons {
     struct GlacisRoute {
         uint256 fromChainId; // 0 means any chain
         bytes32 fromAddress; // 0x00 means any address
-        uint8 fromGmpId; // 0 means any GMP
+        uint160 fromGmpId; // 0 means any GMP, can also hold address
     }
 
     /// @notice De-serialize a uint8 bitmap to an uint8 array
